# openapi/ga4.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GA4 Agent API",
    "version": "1.0.0",
    "description": "API para listar propriedades GA4 e executar relatórios.",
    "license": {
      "name": "Internal Use Only",
      "identifier": "Proprietary"
    }
  },
  "servers": [
    { "url": "https://agente-ga4-api-860407662159.us-central1.run.app" }
  ],
  "paths": {
    "/ga4/properties": {
      "get": {
        "operationId": "findGa4Properties",
        "summary": "Busca propriedades GA4 pelo nome",
        "security": [ { "internalToken": [] } ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Trecho do displayName do property."
          },
          {
            "name": "account",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opcional: conta completa (accounts/123) ou apenas o ID."
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200 },
            "description": "Itens por página (default 50)."
          },
          {
            "name": "pageToken",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Token de paginação retornado pela chamada anterior."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "properties": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Ga4Property" }
                    },
                    "nextPageToken": { "type": "string" }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "ok": true,
                  "properties": [
                    {
                      "name": "properties/270511251",
                      "propertyId": "270511251",
                      "displayName": "Nanda Cora - Real",
                      "parent": "accounts/195680050"
                    }
                  ],
                  "nextPageToken": ""
                }
              }
            }
          },
          "400": { "description": "Requisição inválida para a API GA4" },
          "401": { "description": "Autenticação inválida ou ausente" }
        }
      }
    },
    "/ga4/properties/{propertyId}/metadata": {
      "get": {
        "operationId": "getGa4Metadata",
        "summary": "Lista dimensões e métricas disponíveis",
        "security": [ { "internalToken": [] } ],
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID da propriedade (com ou sem prefixo properties/)."
          },
          {
            "name": "languageCode",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Idioma opcional (ex.: pt-BR)."
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filtra dimensões/métricas cujo nome ou descrição contenham o termo (case-insensitive)."
          },
          {
            "name": "includeDimensions",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Lista separada por vírgula dos apiNames de dimensões que devem ser mantidas."
          },
          {
            "name": "includeMetrics",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Lista separada por vírgula dos apiNames de métricas que devem ser mantidas."
          },
          {
            "name": "maxItems",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "Limite geral de itens retornados (aplicado tanto a dimensões quanto a métricas)."
          },
          {
            "name": "maxDimensions",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "Limite máximo de dimensões retornadas."
          },
          {
            "name": "maxMetrics",
            "in": "query",
            "schema": { "type": "integer" },
            "description": "Limite máximo de métricas retornadas."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "data": { "type": "object" }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "400": { "description": "Requisição inválida para a API GA4" },
          "401": { "description": "Autenticação inválida ou ausente" }
        }
      }
    },
    "/ga4/properties/{propertyId}/runReport": {
      "post": {
        "operationId": "runGa4Report",
        "summary": "Executa relatório GA4",
        "security": [ { "internalToken": [] } ],
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID da propriedade (com ou sem prefixo properties/)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Ga4ReportRequest" },
              "example": {
                "dimensions": [ { "name": "date" } ],
                "metrics": [ { "name": "sessions" }, { "name": "totalUsers" } ],
                "dateRanges": [ { "startDate": "2025-01-01", "endDate": "2025-01-07" } ],
                "limit": 100
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "data": { "type": "object" }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "400": { "description": "Requisição inválida para a API GA4" },
          "401": { "description": "Autenticação inválida ou ausente" }
        }
      }
    }
    ,
    "/ga4/properties/{propertyId}/runPivotReport": {
      "post": {
        "operationId": "runGa4PivotReport",
        "summary": "Executa relatório GA4 com pivôs",
        "security": [ { "internalToken": [] } ],
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID da propriedade (com ou sem prefixo properties/)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Ga4PivotReportRequest" },
              "example": {
                "dimensions": [ { "name": "date" } ],
                "metrics": [ { "name": "sessions" } ],
                "pivots": [
                  {
                    "fieldNames": [ "date" ],
                    "limit": 10,
                    "orderBys": [ { "dimension": { "dimensionName": "date" }, "desc": true } ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ga4GenericResponse" }
              }
            }
          }
        }
      }
    },
    "/ga4/properties/{propertyId}/batchRunReports": {
      "post": {
        "operationId": "batchRunGa4Reports",
        "summary": "Executa múltiplos relatórios GA4 em lote",
        "security": [ { "internalToken": [] } ],
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID da propriedade (com ou sem prefixo properties/)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Ga4CompatibilityRequest" },
              "example": {
                "dimensions": [ { "name": "date" } ],
                "metrics": [ { "name": "sessions" }, { "name": "totalUsers" } ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ga4GenericResponse" }
              }
            }
          }
        }
      }
    },
    "/ga4/properties/{propertyId}/batchRunPivotReports": {
      "post": {
        "operationId": "batchRunGa4PivotReports",
        "summary": "Executa múltiplos relatórios GA4 com pivôs em lote",
        "security": [ { "internalToken": [] } ],
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID da propriedade (com ou sem prefixo properties/)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Ga4BatchRunPivotReportsRequest" },
              "example": {
                "requests": [
                  {
                    "dimensions": [ { "name": "date" } ],
                    "metrics": [ { "name": "sessions" } ],
                    "pivots": [
                      {
                        "fieldNames": [ "date" ],
                        "limit": 7
                      }
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ga4GenericResponse" }
              }
            }
          }
        }
      }
    },
    "/ga4/properties/{propertyId}/runRealtimeReport": {
      "post": {
        "operationId": "runGa4RealtimeReport",
        "summary": "Executa relatório em tempo real",
        "security": [ { "internalToken": [] } ],
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID da propriedade (com ou sem prefixo properties/)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Ga4RealtimeReportRequest" },
              "example": {
                "dimensions": [ { "name": "unifiedScreenName" } ],
                "metrics": [ { "name": "screenPageViews" } ],
                "limit": 10
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ga4GenericResponse" }
              }
            }
          }
        }
      }
    },
    "/ga4/properties/{propertyId}/checkCompatibility": {
      "post": {
        "operationId": "checkGa4Compatibility",
        "summary": "Verifica compatibilidade de dimensões/métricas",
        "security": [ { "internalToken": [] } ],
        "parameters": [
          {
            "name": "propertyId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID da propriedade (com ou sem prefixo properties/)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Ga4CompatibilityRequest" },
              "example": {
                "compatibilityRequest": {
                  "dimensions": [ { "name": "date" } ],
                  "metrics": [ { "name": "sessions" }, { "name": "totalUsers" } ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Ga4GenericResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "internalToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-token",
        "description": "Informe o token interno configurado no Cloud Run."
      }
    },
    "schemas": {
      "Ga4Property": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "propertyId": { "type": "string" },
          "displayName": { "type": "string" },
          "parent": { "type": "string" },
          "timeZone": { "type": "string" },
          "currencyCode": { "type": "string" }
        },
        "additionalProperties": true
      },
      "Ga4ReportRequest": {
        "type": "object",
        "properties": {
          "dimensions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" }
              },
              "required": [ "name" ],
              "additionalProperties": false
            }
          },
          "metrics": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" }
              },
              "required": [ "name" ],
              "additionalProperties": false
            }
          },
          "dateRanges": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "startDate": { "type": "string" },
                "endDate": { "type": "string" }
              },
              "required": [ "startDate", "endDate" ],
              "additionalProperties": false
            }
          },
          "limit": { "type": "integer" }
        },
        "required": [ "metrics", "dateRanges" ],
        "additionalProperties": true
      },
      "Ga4PivotReportRequest": {
        "type": "object",
        "properties": {
          "dimensions": { "type": "array", "items": { "type": "object" } },
          "metrics": { "type": "array", "items": { "type": "object" } },
          "pivots": { "type": "array", "items": { "type": "object" } },
          "dateRanges": { "type": "array", "items": { "type": "object" } }
        },
        "additionalProperties": true
      },
      "Ga4BatchRunReportsRequest": {
        "type": "object",
        "properties": {
          "requests": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Ga4ReportRequest" }
          }
        },
        "required": [ "requests" ],
        "additionalProperties": false
      },
      "Ga4BatchRunPivotReportsRequest": {
        "type": "object",
        "properties": {
          "requests": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Ga4PivotReportRequest" }
          }
        },
        "required": [ "requests" ],
        "additionalProperties": false
      },
      "Ga4RealtimeReportRequest": {
        "type": "object",
        "properties": {
          "dimensions": { "type": "array", "items": { "type": "object" } },
          "metrics": { "type": "array", "items": { "type": "object" } },
          "limit": { "type": "integer" },
          "dimensionFilter": { "type": "object" },
          "metricFilter": { "type": "object" }
        },
        "additionalProperties": true
      },
      "Ga4CompatibilityRequest": {
        "type": "object",
        "properties": {
          "dimensions": { "type": "array", "items": { "type": "object" } },
          "metrics": { "type": "array", "items": { "type": "object" } },
          "dimensionFilter": { "type": "object" },
          "metricFilter": { "type": "object" },
          "compatibilityFilter": {
            "type": "string",
            "description": "Opcional: restringe a resposta para a compatibilidade desejada (ex.: COMPATIBLE)."
          }
        },
        "required": [ "dimensions", "metrics" ],
        "additionalProperties": true
      },
      "Ga4GenericResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "data": { "type": "object" }
        },
        "additionalProperties": true
      }
    }
  }
}